Fixed: crash on iOS 10.

+++ ios/Classes/SwiftFlutterLoginFacebookPlugin.swift
@@ -239,7 +239,7 @@ public class SwiftFlutterLoginFacebookPlugin: NSObject, FlutterPlugin {
         return [
             "token": token.tokenString,
             "userId": token.userID,
-            "expires": Int((token.expirationDate.timeIntervalSince1970 * 1000.0).rounded()),
+            "expires": Int64((token.expirationDate.timeIntervalSince1970 * 1000.0).rounded()),
             "permissions": token.permissions.map {item in item.name},
             "declinedPermissions": token.declinedPermissions.map {item in item.name},
         ]
